Refactor: Move common nominal functions to utils
